+++ components/Transaction.tsx
@@ -91,23 +91,26 @@ const Transaction: React.FC<Props> = ({ data, close }) => {
 
   const handleDelete = async (event: any) => {
     event.preventDefault()
-    setLoading(true)
 
-    try {
-      const res = await supabase
-        .from('transactions')
-        .delete()
-        .eq('id', data?.id)
+    if (window.confirm("Do you really want to delete this?")) {
+      setLoading(true)
 
-      if (res.error) {
-        console.error(res.error)
-        throw new Error()
+      try {
+        const res = await supabase
+          .from('transactions')
+          .delete()
+          .eq('id', data?.id)
+
+        if (res.error) {
+          console.error(res.error)
+          throw new Error()
+        }
+      } catch (e) {
+        console.error(e)
+      } finally {
+        setLoading(false)
+        close()
       }
-    } catch (e) {
-      console.error(e)
-    } finally {
-      setLoading(false)
-      close()
     }
   }
